+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ## Features
  - Feed pages now automatically inject a `<link rel="alternate" type="application/rss+xml">` tag for RSS autodiscovery
+ - Feed index pages can use a `{{ .PostList }}` placeholder that is replaced with the auto-generated, newest-first list of posts
 
 ## Fixes
  -